Seprate controller and routes
