+++ src/routes/digipin.routes.js
@@ -1,30 +1,10 @@
 const express = require('express');
 const router = express.Router();
-const { getDigiPin, getLatLngFromDigiPin } = require('../digipin');
+const digipinController = require('../controllers/digipin.controller');
 
-router.post('/encode', (req, res) => { const { latitude, longitude } = req.body; try { const code = getDigiPin(latitude, longitude); res.json({ digipin: code }); } catch (e) { res.status(400).json({ error: e.message }); } });
+router.post('/encode', digipinController.encodePost);
+router.post('/decode', digipinController.decodePost);
+router.get('/encode', digipinController.encodeGet);
+router.get('/decode', digipinController.decodeGet);
 
-router.post('/decode', (req, res) => { const { digipin } = req.body; try { const coords = getLatLngFromDigiPin(digipin); res.json(coords); } catch (e) { res.status(400).json({ error: e.message }); } });
-
-  router.get('/encode', (req, res) => {
-    const { latitude, longitude } = req.query;
-    try {
-      const code = getDigiPin(parseFloat(latitude), parseFloat(longitude));
-      res.json({ digipin: code });
-    } catch (e) {
-      res.status(400).json({ error: e.message });
-    }
-  });
-  
-  router.get('/decode', (req, res) => {
-    const { digipin } = req.query;
-    try {
-      const coords = getLatLngFromDigiPin(digipin);
-      res.json(coords);
-    } catch (e) {
-      res.status(400).json({ error: e.message });
-    }
-  });
-
-  
 module.exports = router;
